crud operation for user favorite table created

+++ app/src/main/java/com/example/githubusersearcher/data/local/dao/UserDao.kt
@@ -15,7 +15,7 @@ interface UserDao {
     suspend fun insertUser(user: UserEntity)
 
     @Query("DELETE FROM users WHERE userId = :userId")
-    suspend fun deleteUser(userId: Int)
+    suspend fun deleteUser(userId: Long)
 
 
 }
